requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
autoflake==2.3.1
bcrypt==4.2.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
dependency-injector==4.41.0
exceptiongroup==1.2.2
fastapi==0.112.2
fastapi-pagination==0.12.26
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-multipart==0.0.9
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
